refactor(ui): split MacroEditor.tsx into 3 focused modules (610→451 lines)
Extract keycode selection logic and utility functions from MacroEditor.tsx:

- macro-editor-utils.ts (24 lines): parseMacroBuffer, KC_TRNS/KC_NO constants,
  KeycodeAction type, isKeycodeAction guard
- useMacroKeycodeSelection.ts (236 lines): all keycode selection state, popover
  management, click-outside handling, mask mode, revert logic
- MacroEditor.tsx (451 lines): orchestrator with core state, action list handlers,
  save/clear/revert, memory validation, and all JSX rendering
